requirements.txt
customtkinter>=5.2.0
pynput>=1.7.6
pyperclip>=1.8.2
pystray>=0.19.5
Pillow>=10.0.0